Break up machines in to smaller modules
