+++ src/machines/app/constants.ts
@@ -1,0 +1,1 @@
+export const APP_MACHINE_ID = 'app-machine'
