+++ src/machines/app/types.ts
@@ -1,0 +1,12 @@
+import { Player } from 'src/interfaces'
+
+export interface AppStateSchema {
+  states: {
+    creatingCharacter: {}
+    playing: {}
+  }
+}
+
+export interface AppContext {
+  player: Player
+}
