--- src/machines/index.ts
@@ -1,3 +1,0 @@
-export * from './AppMachine'
-export * from './CharacterCreationMachine'
-export * from './PlayAreaMachine'
