+++ src/machines/playArea/index.ts
@@ -1,0 +1,3 @@
+export * from './PlayAreaMachine'
+export * from './types'
+export * from './constants'
